+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres (somewhat) to [Semantic Versioning](https://semver.org/
 ## [Unreleased]
 ### Added
 - Ability to exclude bots from autorole (on by default).
+- `mb?rolestats` command.
 
 
 ## [1.3.2] - 2020-10-06